pyproject.toml
[project]
name = "griptape-nodes-library-standard"
version = "0.67.0"
description = "Default nodes for Griptape Nodes"
authors = [{ name = "Griptape, Inc" }]
requires-python = ">=3.12,<3.13"
dependencies = [
  "griptape[drivers-prompt-amazon-bedrock,drivers-prompt-anthropic,drivers-prompt-cohere,drivers-prompt-ollama,drivers-web-scraper-trafilatura,drivers-web-search-duckduckgo,drivers-web-search-exa,loaders-image,loaders-pdf]>=1.9.3",
  "json-repair>=0.46.1",
  "static-ffmpeg>=2.8",
  "jmespath>=1.0.1",
  "json-schema-to-pydantic>=0.4.6",
  "numpy>=1.26.0",
  "scikit-learn>=1.3.0",
  "threadpoolctl>=3.0.0",
  "pillow>=11.2.1",
  "asteval>=1.0.8",
  "color-matcher",
  "griptape-nodes>=0.77.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["griptape_nodes_library"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # Pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade
]
ignore = [
  "E501", # line too long (handled by formatter)
  "B008", # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = ["pyright>=1.1.396", "pytest-asyncio>=0.21.0", "ruff>=0.8.0", "pytest"]

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["."]
exclude = [
  "**/templates/*",
  ".venv",
  "**/node_modules",
  "**/__pycache__",
  "**/.*",
]
pythonVersion = "3.12"
reportIncompatibleMethodOverride = false
